Show limited access for motor vehicles with dashed green
Many roads are mostly closed to motor vehicles, but allow non-through traffic or special-purpose traffic (e.g. for forestry/farming). These roads tend to be pretty quite and useful for cyclists.

This change represents them by applying the green shading for pure motor-vehicle-free roads in a dashed pattern. This pattern allows the underlying road colour to still show through, which also allows the speed limit on these roads to still be read, for further understanding of how dangerous the limited motor vehicle traffic will be to cyclists.

The green dashes interact fairly well with the dashes used to represent non-smooth road surfaces; the surface dashes are drawn on top of these, but a) they are transparent and b) the surface dashes are spaced further apart.